Refactor and clean up gossip module
